added merge

+++ checkout.py
@@ -14,7 +14,12 @@ import pathing
 import os
 import shutil
 
-def checkout(project_root, repo_directory):
+import merge
+
+def checkout(project_root,
+        repo_directory,
+        merge_commit_id = False,
+        foreign_repo_name = False):
     """TODO: Docstring for checkout.
 
     :project_root: TODO
@@ -32,47 +37,23 @@ def checkout(project_root, repo_directory):
             project_id)
     if os.path.exists(man_file_path):
         man_file = open(man_file_path, 'r')
-        meta_dictionary = process_manifest_file(man_file)
+        meta_dictionary = pathing.read_manifest_file(man_file)
     else:
         # TODO handle this more gracefully.
         print( "Error: couldn't find project \"%\" in repo." % project_id )
         exit(-1)
     copy_files(
-            project_root, repo_directory, meta_dictionary['file_association_list'])
+            project_root,
+            repo_directory,
+            meta_dictionary['file_association_list'],
+            foreign_repo_name,
+            merge_commit_id)
 
-def process_manifest_file(man_file):
-    """TODO: Docstring for parse_header.
-
-    :man_file: TODO
-    :returns: TODO
-
-    """
-    #process_file
-    meta_dictionary = {}
-    line = man_file.readline().rstrip()
-    while line != '=================================':
-        split_line = line.split(' ')
-        meta_dictionary[ split_line[0] ] = split_line[1]
-        line = man_file.readline().rstrip()
-    # now read the comment which might be multi_line
-    #TODO read the comment file
-    line = man_file.readline().rstrip()
-    while line != '=================================':
-        # read file path, hash pairs
-        line = man_file.readline().rstrip()
-        
-
-    # read file association list
-    file_association_list = []
-    line = man_file.readline().rstrip()
-    while len(line) != 0:
-        file_association_list.append(line)
-        line = man_file.readline().rstrip()
-    
-    meta_dictionary['file_association_list'] = file_association_list
-    return meta_dictionary
-
-def copy_files(project_root, repo_directory, file_association_list):
+def copy_files(project_root,
+        repo_directory,
+        file_association_list,
+        foreign_repo_name = False,
+        foreign_merge_commit_id = False):
     project_path_one_down, _ = os.path.split(project_root)
     for association in file_association_list:
         relative_path, repo_hash = os.path.split(association)
@@ -82,5 +63,17 @@ def copy_files(project_root, repo_directory, file_association_list):
         copy_path_one_down, _ = os.path.split(copy_path)
         if not os.path.exists(copy_path_one_down):
             os.makedirs(copy_path_one_down)
-        shutil.copyfile(original_path, copy_path)
+        if( os.path.exists(copy_path)):
+            # this file already exists, so revert to merge behavior if allowed
+            if( foreign_merge_commit_id != False ):
+                merge.merge_file(
+                        project_root,
+                        repo_directory,
+                        copy_path,
+                        original_path,
+                        foreign_repo_name,
+                        foreign_merge_commit_id)
+        else:
+            shutil.copyfile(original_path, copy_path)
         
+
